requirements.txt
pytorch=1.12.1
datasets=2.13.2
transformers=4.24.0
pandas=1.3.5
numpy=1.21.5
scikit-learn=1.0.2
scipy=1.7.3
mlflow
databricks-sdk